Extract sync logic into module
