+++ app/models/concerns/state_country_syncable.rb
@@ -1,0 +1,30 @@
+# frozen_string_literal: true
+
+module StateCountrySyncable
+  extend ActiveSupport::Concern
+
+  included do
+    before_save :sync_state_and_country
+  end
+
+  private
+
+  def sync_state_and_country
+    return unless will_save_change_to_state_code? || will_save_change_to_country_code?
+
+    sync_state
+    sync_country
+  end
+
+  def sync_state
+    return unless state_code.present? && country_code.present?
+
+    self.state_name = ::Carmen::Country.coded(country_code)&.subregions&.coded(state_code)&.name
+  end
+
+  def sync_country
+    return unless country_code.present?
+
+    self.country_name = ::Carmen::Country.coded(country_code)&.name
+  end
+end
