+++ app/models/person.rb
@@ -1,8 +1,7 @@
 # frozen_string_literal: true
 
 class Person < ApplicationRecord
-
-  include Auditable, Concealable, PersonalInfo, Searchable, Subscribable, Matchable, UrlAccessible
+  include Auditable, Concealable, PersonalInfo, Searchable, StateCountrySyncable, Subscribable, Matchable, UrlAccessible
   extend FriendlyId
 
   strip_attributes collapse_spaces: true
@@ -29,8 +28,6 @@ class Person < ApplicationRecord
             format: {with: VALID_EMAIL_REGEX}
   validates :phone, allow_blank: true, format: {with: VALID_PHONE_REGEX}
   validates_with BirthdateValidator
-
-  before_save :sync_state_and_country
 
   # This method needs to extract ids and run a new search to remain compatible
   # with the scope `.with_age_and_effort_count`.
@@ -86,24 +83,5 @@ class Person < ApplicationRecord
 
   def generate_new_topic_resource?
     true
-  end
-
-  def sync_state_and_country
-    return unless will_save_change_to_state_code? || will_save_change_to_country_code?
-
-    sync_state
-    sync_country
-  end
-
-  def sync_state
-    return unless state_code.present? && country_code.present?
-
-    self.state_name = ::Carmen::Country.coded(country_code)&.subregions&.coded(state_code)&.name
-  end
-
-  def sync_country
-    return unless country_code.present?
-
-    self.country_name = ::Carmen::Country.coded(country_code)&.name
   end
 end
